moved base dt class to dt module; removed <raises> entries from new functions doc
