+++ feature_engine/base_transformers.py
@@ -15,11 +15,7 @@ from feature_engine.dataframe_checks import (
     _is_dataframe,
 )
 from feature_engine.validation import _return_tags
-from feature_engine.variable_manipulation import (
-    _convert_variables_to_datetime,
-    _find_or_check_datetime_variables,
-    _find_or_check_numerical_variables,
-)
+from feature_engine.variable_manipulation import _find_or_check_numerical_variables
 
 
 class BaseNumericalTransformer(BaseEstimator, TransformerMixin):
@@ -142,134 +138,6 @@ class BaseNumericalTransformer(BaseEstimator, TransformerMixin):
         # check if dataset contains na or inf
         _check_contains_na(X, self.variables_)
         _check_contains_inf(X, self.variables_)
-
-        return X
-
-    # for the check_estimator tests
-    def _more_tags(self):
-        return _return_tags()
-
-
-class DateTimeBaseTransformer(BaseEstimator, TransformerMixin):
-    """shared set-up procedures across datetime transformers"""
-
-    def _select_variables_from_dict(
-        self, X: pd.DataFrame, user_dict_: Dict
-    ) -> pd.DataFrame:
-        """
-        Checks that input is a dataframe, checks that variables in the dictionary
-        entered by the user are of type datetime.
-
-        Parameters
-        ----------
-        X : Pandas DataFrame
-
-        user_dict_ : Dictionary. Default = None
-            Any dictionary allowed by the transformer and entered by user.
-
-        Raises
-        ------
-        TypeError
-            If the input is not a Pandas DataFrame or a numpy array
-            If any of the variables in the dictionary are not datetime
-        ValueError
-            If there are no numerical variables in the df or the df is empty
-            If the variable(s) contain null values
-
-        Returns
-        -------
-        X : Pandas DataFrame
-            The same dataframe entered as parameter
-        """
-        # check input dataframe
-        X = _is_dataframe(X)
-
-        # find or check for datetime variables
-        variables = [x for x in user_dict_.keys()]
-        self.variables_ = _find_or_check_datetime_variables(X, variables)
-
-        # check if dataset contains na or inf
-        _check_contains_na(X, self.variables_)
-        _check_contains_inf(X, self.variables_)
-
-        return X
-
-    def fit(self, X: pd.DataFrame, y: Optional[pd.Series] = None) -> pd.DataFrame:
-        """
-        Checks that input is a dataframe, finds datetime variables,
-        or alternatively checks that variables entered by the user are
-        or can be converted to type datetime
-
-        Parameters
-        ----------
-        X : Pandas DataFrame
-
-        y : Pandas Series, np.array. Default = None
-            Parameter is necessary for compatibility with sklearn Pipeline.
-
-        Raises
-        ------
-        TypeError
-            If the input is not a Pandas DataFrame or a numpy array
-            If any of the user provided variables are not or cannot be converted
-            to datetime
-        ValueError
-            If there are no datetime variables in the df or the df is empty
-            If the variable(s) contain null values
-
-        Returns
-        -------
-        X : Pandas DataFrame
-            The same dataframe entered as parameter
-        """
-
-        # check input dataframe
-        X = _is_dataframe(X)
-
-        # find or check for datetime variables
-        self.variables_ = _find_or_check_datetime_variables(X, self.variables)
-
-        # check if datetime variables contains na
-        _check_contains_na(X, self.variables_)
-
-        return X
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Checks that the input is a dataframe and of the same size than the one used
-        in the fit method. Checks absence of NA and Inf.
-
-        Parameters
-        ----------
-        X : Pandas DataFrame
-
-        Raises
-        ------
-        TypeError
-            If the input is not a Pandas DataFrame
-        ValueError
-            - If the variable(s) contain null values
-            - If the df has different number of features than the df used in fit()
-
-        Returns
-        -------
-        X : Pandas DataFrame.
-            The same dataframe entered by the user.
-        """
-
-        # Check method fit has been called
-        check_is_fitted(self)
-
-        # check that input is a dataframe
-        X = _is_dataframe(X)
-
-        # Check if input data contains same number of columns as dataframe used to fit.
-        _check_input_matches_training_df(X, self.n_features_in_)
-
-        # check if dataset contains na
-        _check_contains_na(X, self.variables_)
-
-        X = _convert_variables_to_datetime(X, self.variables_, **self.kwargs)
 
         return X
