+++ feature_engine/parameter_checks.py
@@ -50,17 +50,6 @@ def _parse_features_to_extract(feats_requested, feats_supported: List[str]) -> A
     feats_supported: list[str]
         list of supported date features that the transformer is able
         to extract from the existing datetime variables
-
-    Raises
-    ------
-    TypeError
-        If requested features is not a string or list
-        If supported features is not a list of strings
-
-    ValueError
-        If one or more requested features are not supported
-        by the transformer
-
     """
     if not isinstance(feats_requested, (str, list)):
         raise TypeError("feats_requested must be either a str or a list of str")
